fix: the logic of creating a membership has been fixed - now the system does not allow entering a negative duration, and also supports incomplete months for calculating the membership end date

+++ LibraryWeb/Models/MembershipType.cs
@@ -13,7 +13,7 @@ namespace LibraryWeb.Models
 
         [Required]
         [Range(1, 36)]
-        public int DurationMonths { get; set; }
+        public double DurationMonths { get; set; }
 
         [Required]
         [Range(0, 10000)]
